+++ BimServer/src/org/bimserver/database/actions/DownloadProjectsDatabaseAction.java
@@ -92,7 +92,7 @@ public class DownloadProjectsDatabaseAction extends AbstractDownloadDatabaseActi
 							if (totalSize == 0) {
 								setProgress("Preparing download...", 0);
 							} else {
-								setProgress("Preparing download...", Math.round(100L * total.get() / totalSize));
+								setProgress("Preparing download...", (int) Math.round(100.0 * total.get() / totalSize));
 							}
 						}
 					});
